Add basic.py
